Replace :trigger with :table-alias
We don't actually _want_ to be naming database objects nor particularly
concerning ourselves with their names we can let the PL/SQL worry about
those sorts of details and restrict our communication to be about the
table.

However, we do need to provide a sensible alias for the table that can
be used where its full name would be too long. This needs to be
something that we're specifying so that we have a record of it and a
manual means of preventing collisions.